Create percent_tumour_area_from_depth
This is used to identify cores with 0% or greater than 80% tumour for exclusion. Tumour percentage is calculated from the depth of the "leading edge" of the tumour on the core. The leading edge is defined as the axis of the tumour on the core parallel to the contour of the WS tumour.